requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
decorator==5.0.9
distlib==0.3.2
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.1
Flask-API==3.0.post1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
fuzzywuzzy==0.18.0
greenlet==1.1.0
idna==3.2
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
levenshtein==0.12.0
lxml==4.6.3
MarkupSafe==2.0.1
marshmallow==3.12.1
mccabe==0.6.1
mysql==0.0.3
mysql-connector-python==8.0.25
mysqlclient==2.0.3
pipenv==2021.5.29
protobuf==3.17.1
py3-validate-email==1.0.0
pycodestyle==2.7.0
pylint==2.6.0
pyodbc==4.0.31
PySocks==1.7.1
python-decouple==3.4
python-http-client==3.3.2
requests==2.25.1
sendgrid==6.7.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
starkbank-ecdsa==1.1.1
toml==0.10.2
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
wrapt==1.12.1
